Move combineValidDictionaries test to its own file
